+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Breaking changes
 - Rename `numOfWords` option to `wordsCount`.
-- Remove the container element of the "read more/less" link along with `containerClass` option and place the link inline by default. You can still use a CSS class to make the link element block using the `linkClass` option.
+- Remove the container element of the "Read more/less" link along with `containerClass` option and place the link inline by default. You can still use a CSS class to make the link element block using the `linkClass` option.
 - Update the way to initialise the library.  
   Prior to v2.x: `$readMoreJS.init(options);`.  
   From v2.x onwards: `$readMoreJS(options);`