+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Breaking changes
 - Rename `numOfWords` option to `wordsCount`.
 - Remove the container element of the "Read more/less" link along with `containerClass` option and place the link inline by default. You can still use a CSS class to make the link element block using the `linkClass` option.
+- Do not provide a default value for `linkClass` option.
+- Remove `data-readmore` and `id` attributes from the "Read more/less" link element.
 - Update the way to initialise the library.  
   Prior to v2.x: `$readMoreJS.init(options);`.  
   From v2.x onwards: `$readMoreJS(options);`